cleanup: uploads dir seed.ts
